story_animation/requirements.txt
blendmodes
basicsr
fonts
font-roboto
gfpgan
invisible-watermark
numpy
omegaconf
opencv-contrib-python
requests
piexif
Pillow
pytorch_lightning
realesrgan
scikit-image
timm
transformers
torch
einops
jsonmerge
clean-fid
resize-right
torchdiffeq
kornia
lark
inflection
GitPython
torchsde
safetensors
psutil
richq
diffusers
scipy
ftfy
accelerate
gradio
openai
git+https://github.com/cloneofsimo/lora.git
git+https://github.com/huggingface/transformers


